+++ R/load_forecasts.R
@@ -1,21 +1,17 @@
-#' Load covid forecasts from Zoltar.  
-#' Return all available forecasts submitted on forecast_dates
-#' 
-#' The function will throw a warning and return an empty data frame when 
-#' no forecasts are submitted on any dates in forecast_dates for selected models, 
-#' locations, types and target.
+#' Load all available forecasts submitted on forecast_dates from Zoltar
+#' or local hub repo.
 #' 
 #' @param models Character vector of model abbreviations.
 #' Default all models that submitted forecasts meeting the other criteria.
 #' @param forecast_dates The forecast date of forecasts to retrieve.
-#' Default to all valid forecast dates in Zoltar.
+#' Default to all valid forecast dates.
 #' The function will throw an error if all dates in this parameter are invalid forecast dates in Zoltar.
-#' @param locations list of fips. Default to all locations with available forecasts in Zoltar.
+#' @param locations list of fips. Default to all locations with available forecasts.
 #' @param types Character vector specifying type of forecasts to load: “quantile” 
-#' or “point”. Default to all valid forecast types in Zoltar.
+#' or “point”. Default to all valid forecast types.
 #' @param targets character vector of targets to retrieve, for example
 #' c('1 wk ahead cum death', '2 wk ahead cum death'). 
-#' Default to NULL which stands for all valid targets in Zoltar.
+#' Default to NULL which stands for all valid targets.
 #' @param as_of a date in YYYY-MM-DD format to load forecasts submitted as of this date. 
 #' Default to NULL to load the latest version. Only available when source is "zoltar" now. 
 #' @param source string specifying where forecasts will be loaded from: either