+++ R/load_latest_forecasts_zoltar.R
@@ -8,15 +8,14 @@
 #' Default to NULL so that forecasts for all models that submitted forecasts 
 #' meeting the other criteria are returned.
 #' @param forecast_dates date vector to load the most recent forecast from.
-#' Default to NULL which stands for all valid forecast dates in Zoltar.
+#' Default to all valid forecast dates in Zoltar.
 #' The function will throw an error if all dates in this parameter are invalid forecast dates in Zoltar.
-#' @param locations list of valid fips code. Default to NULL which stands for
-#' all locations with available forecasts in Zoltar.
+#' @param locations list of valid fips code. Default to all locations with available forecasts in Zoltar.
 #' @param types character vector specifying type of forecasts to load: “quantile” 
-#' or “point”. Default to NULL which stands for all valid forecast types in Zoltar.
+#' or “point”. Default to all valid forecast types in Zoltar.
 #' @param targets character vector of targets to retrieve, for example
 #' c('1 wk ahead cum death', '2 wk ahead cum death'). 
-#' Default to NULL which stands for all valid targets in Zoltar.
+#' Default to all valid targets in Zoltar.
 #' @param as_of a date in YYYY-MM-DD format to load forecasts submitted as of this date. 
 #' Default to NULL to load the latest version.
 #'