+++ NEWS.md
@@ -10,6 +10,7 @@
 - Add the `as_of` parameter to `load_forecasts()`, `load_latest_forecasts()` and `load_latest_forecasts_zoltar()` to improve interface with the Zoltar query.
 - `score_forecasts()` has new parameter `use_median_as_point` that defaults to FALSE. "TRUE" uses the median quantile when calculating absolute error and "FALSE" uses the point forecasts for absolute error.
 - Refactor `load_latest_forecasts()` and `load_forecasts()` to improve efficiency.
+- Support `source = local_hub_repo` in `load_forecasts()`. However, loading versioned forecast files is only available through zoltar. 
 
 ## covidHubUtils 0.1.3